Knight ability and base of Testing Frame
Completed the knight ability (for now). Also the Testing Frame has been
started. This will be useful for testing the game.

+++ game/unit/properties/HealthProperty.java
@@ -18,7 +18,7 @@ public class HealthProperty extends Property<Integer> {
     }
 
     public double percentageHealth() {
-	return getCurrentPropertyValue() / defaultPropValue;
+	return (double) getCurrentPropertyValue() / defaultPropValue;
     }
 
     public void takeDamage(Damage damage) {
